backend/requirements.txt
Django>=4.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
scikit-learn>=1.0.0
numpy>=1.21.0

